This commit introduces sensor identification by ProductID from packets, in the hdl_grabber_driver.cpp
src/hdl_grabber_nodelet.cpp has also changes to speedup code and avoid possible issues.
some functions are commented. these changes do not compromise (or shoudn't) how the
pointclouds are provided. also the launcher uses a standalone nodelet, that was used
to profile the cpp.

these lines are clearly identificable by: // AUGUSTO DELETES THESE LINES sections.

launchers are changed to simplify and avoid parameters that are not used with the Citroen C4.